requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
pydantic-core>=2.27.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pillow>=11.0.0
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
weaviate-client>=4.11.0
email-validator>=2.2.0
python-multipart>=0.0.5
openai>=1.68.0
autogen>=0.8.0
pytest>=8.3.0
httpx>=0.28.0
aiosqlite>=0.20.0
asyncpg>=0.29.0 